refactor: grouping the utils.go(requests_sender.go) and errors(custom_errors.go) in one utils folder for more calrity and structure organization
