+++ pkg/chargily/utils/custom_errors.go
@@ -1,4 +1,4 @@
-package chargily
+package utils
 
 import "errors"
